PCIe: Added few additional checks and fixes
- The return type of function val_pcie_get_bdf_config_addr()
  is changed to uint64 from uint32.
- BDF table entry to ignore Host/PCI bridge
- Bit field check to only the required bitfield for RSVDP
- Added fix in incrementing to the next bar offset
- Added addition seg check to get the downstream function
- Update the prints to reflect the return types for FAR
  and ESR
- Skipped host Bridge in being added to the bdf table
- Segment check in test_p044
- Skipped legacy PCI device being added to the bdf table

+++ val/include/sbsa_avs_pcie.h
@@ -138,7 +138,7 @@ typedef struct {
 void     val_pcie_write_cfg(uint32_t bdf, uint32_t offset, uint32_t data);
 uint32_t val_pcie_read_cfg(uint32_t bdf, uint32_t offset, uint32_t *data);
 uint32_t val_get_msi_vectors (uint32_t bdf, PERIPHERAL_VECTOR_LIST **mvector);
-uint32_t val_pcie_get_bdf_config_addr(uint32_t bdf);
+uint64_t val_pcie_get_bdf_config_addr(uint32_t bdf);
 
 
 typedef enum {
